refactor: move config-diff compare utils, types, services to common-lib
