+++ src/Pages/Applications/DevtronApps/Details/AppConfigurations/Navigation/EnvironmentOverrideRouter.tsx
@@ -17,6 +17,7 @@
 import React, { useState, useEffect, Fragment } from 'react'
 import { useParams, useLocation, useRouteMatch, NavLink, Link } from 'react-router-dom'
 import {
+    EnvResourceType,
     showError,
     DeleteDialog,
     ConfirmationDialog,
@@ -37,7 +38,7 @@ import { ReactComponent as More } from '../../../../../../assets/icons/ic-more-o
 import { ReactComponent as DeleteIcon } from '../../../../../../assets/icons/ic-delete-interactive.svg'
 import { ReactComponent as ProtectedIcon } from '../../../../../../assets/icons/ic-shield-protect-fill.svg'
 import warn from '../../../../../../assets/icons/ic-warning.svg'
-import { EnvResourceType, JobEnvOverrideRouteProps } from '../AppConfig.types'
+import { JobEnvOverrideRouteProps } from '../AppConfig.types'
 import { groupHeading } from '../../../../../../components/CIPipelineN/Constants'
 import { RESOURCE_ACTION_MENU } from '../../../../../../components/ResourceBrowser/Constants'
 import { groupStyle } from '../../../../../../components/v2/common/ReactSelect.utils'
